record video walkthrough for app

+++ app/controllers/reviews_controller.rb
@@ -15,38 +15,38 @@ class ReviewsController < ApplicationController
       render json: review, status: :created
   end
 
-  # def update
-  #   review = Review.find(params[:id])
-  #   if review && review.user_id == @current_user.id
-  #     review.update!(review_params)
-  #     render json: review , status: :created
-  #   else
-  #     render json: "Invalid Credentials", status: :unauthorized
-  #   end
-  # end
-
   def update
-      review = Review.find(params[:id])
+    review = Review.find(params[:id])
+    if review && review.user_id == @current_user.id
       review.update!(review_params)
       render json: review , status: :created
+    else
+      render json: "Invalid Credentials", status: :unauthorized
+    end
   end
 
-  def destroy
-    review = Review.find(params[:id])
-    review.destroy
-    head :no_content
-  end
+  # def update
+  #     review = Review.find(params[:id])
+  #     review.update!(review_params)
+  #     render json: review , status: :created
+  # end
 
   # def destroy
   #   review = Review.find(params[:id])
-  #   if review && review.user_id == @current_user.id
-  #     # byebug
-  #     review.destroy
-  #     head :no_content
-  #   else
-  #     render json: "Invalid Credentials", status: :unauthorized
-  #   end
+  #   review.destroy
+  #   head :no_content
   # end
+
+  def destroy
+    review = Review.find(params[:id])
+    if review && review.user_id == @current_user.id
+      # byebug
+      review.destroy
+      head :no_content
+    else
+      render json: "Invalid Credentials", status: :unauthorized
+    end
+  end
 
   private
